+++ cs336_basics/modules/__init__.py
@@ -1,11 +1,12 @@
 from .embedding import Embedding
-from .rmsnorm import RMSNorm
+from .norm import RMSNorm
 from .linear import Linear
 from .ffn import SwiGLU
 from .rope import RotaryPositionalEmbedding
-from .attention import softmax, scaled_dot_product_attention, MultiHeadSelfAttention
+from .attention import scaled_dot_product_attention, MultiHeadSelfAttention
 from .transformer_block import TransformerBlock
 from .lm import TransformerLM
+from .loss import cross_entropy
 
 __all__ = [
     "Embedding",
@@ -13,9 +14,9 @@ __all__ = [
     "Linear",
     "SwiGLU",
     "RotaryPositionalEmbedding",
-    "softmax",
     "scaled_dot_product_attention",
     "MultiHeadSelfAttention",
     "TransformerBlock",
     "TransformerLM",
+    "cross_entropy",
 ]
