+++ cs336_basics/modules/attention.py
@@ -5,6 +5,8 @@ import torch.nn as nn
 from einops import einsum, rearrange
 from jaxtyping import Float, Bool, Int
 
+from cs336_basics.functional import softmax
+
 from .linear import Linear
 from .rope import RotaryPositionalEmbedding
 
@@ -122,19 +124,3 @@ def scaled_dot_product_attention(
     return einsum(
         attn_weights, value, "... seq_len_q seq_len_k, ... seq_len_k d_v -> ... seq_len_q d_v"
     )
-
-
-def softmax(x: Float[Tensor, " ..."], dim: int) -> Float[Tensor, " ..."]:
-    """
-    Computes the softmax of the input tensor along the specified dimension.
-
-    Args:
-        x (torch.Tensor): Input tensor.
-        dim (int): Dimension along which to compute the softmax.
-
-    Returns:
-        torch.Tensor: Softmax of the input tensor.
-    """
-    max_values, _ = torch.max(x, dim=dim, keepdim=True)
-    x_exp = torch.exp(x - max_values)  # for numerical stability
-    return x_exp / torch.sum(x_exp, dim=dim, keepdim=True)
